requirements-web.txt
# Phone Agent Web Interface Dependencies
# Install with: pip install -r requirements-web.txt

# Core Flask dependencies
Flask==2.3.3
Flask-SocketIO==5.3.6

# WebSocket support
python-socketio==5.9.0
eventlet==0.33.3

# HTTP requests (for model service checks)
requests==2.31.0

# Additional utilities
python-engineio==4.7.1
bidict==0.22.1